requirements.txt
absl-py==0.11.0
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
cffi 
chardet==4.0.0
click==7.1.2
confuse==1.0.0
cycler==0.10.0
Flask==1.1.2
gast==0.3.3
gevent 
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet 
grpcio==1.34.0
h5py==2.10.0
htmlmin==0.1.12
idna==2.10
importlib-metadata==3.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib 
kaggle==1.5.6
Keras-Preprocessing==1.1.2
kiwisolver
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe 
matplotlib
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy 
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.1.2.30
opt-einsum==3.3.0
pandas 
pandas-profiling==2.5.0
phik==0.9.9
Pillow 
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser 
pylint==2.4.4
pyparsing==2.4.7
pytest-pylint==0.15.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz 
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.0
scipy 
seaborn==0.11.1
six 
tangled-up-in-unicode==0.0.3
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl 
tornado 
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.26.2
visions==0.2.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface 